+++ port/gamepadmotionhelper/GamepadMotion.cpp
