+++ port/gamepadmotionhelper/GamepadMotion.h
